common fluidgame object
